// miPrimeraApp/src/interface/Post.d.ts
 
 
  export  interface GET_POST{
  userId:number,
  id:number;
  title:string;
  body:string

 }

 export  interface CREATE_POST{
    
    title:string;
    body:string
    userId:number,
  
   }
  

 /**
  
  {
    "userId": 1,
    "id": 1,
    "title": "sunt aut facere repellat provident occaecati excepturi optio reprehenderit",
    "body": "quia et suscipit\nsuscipit recusandae consequuntur expedita et cum\nreprehenderit molestiae ut ut quas totam\nnostrum rerum est autem sunt rem eveniet architecto"
  },

  */